+++ src/main/java/com/range/rpms/common/config/SwaggerConfig.java
@@ -17,7 +17,7 @@ public class SwaggerConfig {
         return new OpenAPI()
                 .info(new Info().title("Range Package Manager Server")
                         .description("A Server for managing packages")
-                        .version("v2.0.0")
+                        .version("v2.7.0")
                         .license(new License().name("GPL 3.0").url("https://github.com/range79/rpms-server/blob/main/LICENSE")))
                 .externalDocs(new ExternalDocumentation()
                         .description("Github")