+++ src/main/java/com/range/rpms/common/util/UserContext.java
@@ -1,6 +1,5 @@
 package com.range.rpms.common.util;
 
 public interface UserContext {
-
     long getCurrentUserId();
 }