+++ src/main/java/com/range/rpms/error/dto/ErrorResponse.java
@@ -1,0 +1,19 @@
+package com.range.rpms.error.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.springframework.stereotype.Component;
+
+import java.time.LocalDateTime;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class ErrorResponse {
+    private String message;
+    private String detail;
+    private LocalDateTime timestamp;
+}