+++ src/main/java/com/range/rpms/error/service/impl/ErrorServiceImpl.java
@@ -1,0 +1,41 @@
+package com.range.rpms.error.service.impl;
+
+import com.range.rpms.error.dto.ErrorResponse;
+import com.range.rpms.error.domain.model.ErrorTypes;
+import com.range.rpms.error.domain.model.Errors;
+import com.range.rpms.error.domain.repository.ErrorRepository;
+import com.range.rpms.error.service.ErrorService;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.stereotype.Component;
+
+@Component
+public class ErrorServiceImpl implements ErrorService {
+    private ErrorRepository errorRepository;
+    public ErrorServiceImpl(ErrorRepository errorRepository) {
+        this.errorRepository = errorRepository;
+    }
+    @Override
+    public ErrorResponse saveError(Errors errors) {
+      Errors error =    errorRepository.save(errors);
+   return ErrorResponse.builder()
+           .detail(error.getDetail())
+           .message(error.getMessage())
+           .timestamp(error.getTimestamp()).build();
+    }
+
+    @Override
+    public Page<Errors> findAll(int size, int page) {
+     return errorRepository.findAll(PageRequest.of(page, size));
+    }
+
+    @Override
+    public Page<Errors> findServerErrors(int size, int page) {
+        return errorRepository.findAllByErrorType(ErrorTypes.SERVER_ERROR, PageRequest.of(page, size));
+    }
+
+    @Override
+    public Page<Errors> findClientErrors(int size, int page) {
+        return errorRepository.findAllByErrorType(ErrorTypes.CLIENT_ERROR, PageRequest.of(page, size));
+    }
+}