+++ src/main/java/com/range/rpms/likes/domain/repository/PackageLikeRepository.java
@@ -2,6 +2,7 @@ package com.range.rpms.likes.domain.repository;
 
 import com.range.rpms.likes.domain.entity.PackageLike;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;